email and password beefed up in database, synch and seed seperated for readibilty, random clean up done
