+++ server/db/models/Game.js
@@ -1,10 +1,8 @@
 const db = require('../db')
 const Sequelize = require('sequelize')
-const { STRING, FLOAT, INTEGER, ENUM, BOOLEAN, DATE, BIGINT } = Sequelize;
+const { STRING, INTEGER, BOOLEAN, DATE, BIGINT } = Sequelize;
 
 const Game = db.define('game', {
-    //Once matching does not need to be exact time/date, and location will need 
-    //to be put back 
     winner: { 
       type: STRING, 
     },  
@@ -23,11 +21,8 @@ const Game = db.define('game', {
     // this will need to be changed starting off really restirctive to  make testing
     // easier will need to replaced with LAT, LONG I assume
     location: {  
-      type: STRING
-      // type: ENUM('COURT 1', 'COURT 2', 'COURT 3','COURT 4','COURT 5'), 
+      type: STRING 
     },
-    // I think date and time can go together, keeping it very simple at the moment
-    // just time date will be added back again very restrive for testing 
     time: { 
       type: BIGINT
     },
