--- src/env.js
@@ -1,3 +1,0 @@
-module.exports = {
-  apiKey: 'AIzaSyCVN7I_2Ha7s8fF2oR1-NxI58Dc57SO6fA'
-};
